Add Client Functionality

+++ src/contexts/SalesContext.jsx
@@ -3,7 +3,6 @@ import {
   getSalesFromFirestore,
   addSaleToFirestore,
 } from "../services/firebaseService";
-
 import { v4 as uuidv4 } from "uuid";
 
 export const SalesContext = createContext();
@@ -24,19 +23,19 @@ export const SalesProvider = ({ children }) => {
     const syncOfflineSales = async () => {
       const localSales = JSON.parse(localStorage.getItem("offlineSales")) || [];
       if (localSales.length > 0) {
-        const salesData = await getSalesFromFirestore(); // Obtén las ventas actuales del servidor
-        const existingSaleIds = new Set(salesData.map((sale) => sale.id)); // Crea un conjunto de IDs de ventas existentes
+        const salesData = await getSalesFromFirestore();
+        const existingSaleIds = new Set(salesData.map((sale) => sale.id));
 
         const salesToSync = localSales.filter(
           (sale) => !existingSaleIds.has(sale.id)
-        ); // Filtra las ventas que no están ya en el servidor
+        );
 
         for (let sale of salesToSync) {
           await addSaleToFirestore(sale);
         }
 
         localStorage.removeItem("offlineSales");
-        setSales([...salesData, ...salesToSync]); // Actualiza el estado con las ventas sincronizadas y las del servidor
+        setSales([...salesData, ...salesToSync]);
       }
     };
 
@@ -52,21 +51,53 @@ export const SalesProvider = ({ children }) => {
     const saleId = uuidv4();
     const saleWithId = { ...newSale, UUID: saleId, date: tempTimestamp };
 
-    if (navigator.onLine) {
-      try {
-        const saleFromServer = await addSaleToFirestore(saleWithId);
-        setSales((prevSales) => [...prevSales, saleFromServer]);
-      } catch (error) {
-        console.error("Error adding sale:", error);
+    const today = new Date().toLocaleDateString();
+
+    const existingSaleIndex = sales.findIndex(
+      (sale) =>
+        sale.cliente === newSale.cliente && new Date(sale.date).toLocaleDateString() === today
+    );
+
+    if (existingSaleIndex !== -1) {
+      const updatedSales = [...sales];
+      const existingSale = updatedSales[existingSaleIndex];
+      existingSale.items = [
+        ...existingSale.items,
+        ...newSale.items,
+      ];
+      existingSale.importe += newSale.importe;
+
+      if (navigator.onLine) {
+        try {
+          await addSaleToFirestore(existingSale);
+        } catch (error) {
+          console.error("Error updating sale:", error);
+        }
+      } else {
+        localStorage.setItem(
+          "offlineSales",
+          JSON.stringify([...updatedSales])
+        );
       }
+
+      setSales(updatedSales);
     } else {
-      const offlineSales =
-        JSON.parse(localStorage.getItem("offlineSales")) || [];
-      localStorage.setItem(
-        "offlineSales",
-        JSON.stringify([...offlineSales, saleWithId])
-      );
-      setSales((prevSales) => [...prevSales, saleWithId]);
+      if (navigator.onLine) {
+        try {
+          const saleFromServer = await addSaleToFirestore(saleWithId);
+          setSales((prevSales) => [...prevSales, saleFromServer]);
+        } catch (error) {
+          console.error("Error adding sale:", error);
+        }
+      } else {
+        const offlineSales =
+          JSON.parse(localStorage.getItem("offlineSales")) || [];
+        localStorage.setItem(
+          "offlineSales",
+          JSON.stringify([...offlineSales, saleWithId])
+        );
+        setSales((prevSales) => [...prevSales, saleWithId]);
+      }
     }
   };
